Week3 sanya zol (#34)
* fix lint

* upd AllExceptionFilter, upd Swagger, add add/remove favoriteSkill

* fix lint

* fix allExceptionFilter, fix add Skill

* fix

---------

+++ src/skills/skills.service.ts
@@ -1,8 +1,10 @@
 import {
+  BadRequestException,
   ForbiddenException,
   Injectable,
   NotFoundException,
 } from '@nestjs/common';
+import { validate as isUuid } from 'uuid';
 import { CreateSkillDto } from './dto/create-skill.dto';
 import { UpdateSkillDto } from './dto/update-skill.dto';
 import { unlink } from 'node:fs';
@@ -10,10 +12,12 @@ import * as path from 'path';
 import { InjectRepository } from '@nestjs/typeorm';
 import { Repository } from 'typeorm';
 import { Skill } from './entities/skill.entity';
+import { User } from 'src/users/entities/users.entity';
 
 @Injectable()
 export class SkillsService {
   constructor(
+    @InjectRepository(User) private userRepository: Repository<User>,
     @InjectRepository(Skill) private skillRepository: Repository<Skill>,
   ) {}
 
@@ -21,6 +25,7 @@ export class SkillsService {
     return await this.skillRepository.save({
       ...createSkillDto,
       owner: { id: userId },
+      message: 'Навык создан',
     });
   }
 
@@ -37,12 +42,13 @@ export class SkillsService {
     const [skills, total] = await qb
       .skip((page - 1) * limit)
       .take(limit)
+      .leftJoinAndSelect('skill.owner', 'owner')
       .getManyAndCount();
 
     const totalPages = Math.ceil(total / limit);
 
     if (page > totalPages && totalPages !== 0) {
-      throw new NotFoundException('Page not found');
+      throw new NotFoundException('Страница не найдена');
     }
 
     return {
@@ -52,33 +58,46 @@ export class SkillsService {
     };
   }
 
-  async update(userId: string, id: string, updateSkillDto: UpdateSkillDto) {
-    const skill = await this.userIsOwner(id, userId);
+  async update(
+    userId: string,
+    skillId: string,
+    updateSkillDto: UpdateSkillDto,
+  ) {
+    await this.userIsOwner(skillId, userId);
     return await this.skillRepository.save({
-      ...skill,
       ...updateSkillDto,
+      message: `Навык с id ${skillId} обновлен у пользователя`,
     });
   }
 
-  async remove(id: string, userId: string) {
-    const skill = await this.userIsOwner(id, userId);
+  async remove(skillId: string, userId: string) {
+    if (!isUuid(skillId)) {
+      throw new BadRequestException('Некорректный UUID навыка');
+    }
+    const skill = await this.userIsOwner(skillId, userId);
     skill.images.forEach((image) => {
       const relativePath = image.startsWith('/') ? image.slice(1) : image;
       const absolutePath = path.join(process.cwd(), relativePath);
       unlink(absolutePath, (err) => {
         if (err) {
-          console.error(`Ошибка при удалении файла ${absolutePath}:`, err);
+          throw new BadRequestException('err');
         }
       });
     });
-    await this.skillRepository.delete(id);
-    return `Skill with id ${id} has been removed`;
+    await this.skillRepository.delete(skillId);
+    return { message: `Навык с id ${skillId} удалён у пользователя` };
   }
 
-  async userIsOwner(id: string, userId: string) {
-    const skill = await this.skillRepository.findOneByOrFail({ id });
+  async userIsOwner(skillId: string, userId: string) {
+    const skill = await this.skillRepository.findOne({
+      where: { id: skillId },
+      relations: ['owner'],
+    });
+    if (!skill) throw new NotFoundException('Навык не найден');
     if (skill.owner.id !== userId) {
-      throw new ForbiddenException('Недостаточно прав');
+      throw new ForbiddenException(
+        `Пользователь ${userId} пытается обновить навык ${skillId}, которым не владеет`,
+      );
     }
     return skill;
   }
